Move canvas related code to a separate package
